Refactoring file handing based on 'Confident Ruby'...
Replace hacky initialize file handling code with a checked_file_for_open() method which returns a @file variable
which can be safely passed to the Nokogiri/open handling. If an unreadable file, then we raise an error.
Able to remove much of the old initialize code entirely and better show the intent of the analyze method.